requirements.txt
accelerate==0.29.2
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.6
Jinja2==3.1.3
lxml==5.2.1
magic-filter==1.0.12
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
packaging==24.0
psutil==5.9.8
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pyTelegramBotAPI==4.17.0
python-docx==1.1.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
soundfile==0.12.1
SpeechRecognition==3.10.3
sympy==1.12
telebot==0.0.5
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
